Explicitly thread lifetimes through `Document::root_element`
This patch ensures that the `Node` produced by
`Document::root_element` matches the lifetimes of the producing
`Document`. Prior to this patch, Rust’s lifetime-assumption model
was applying the implicit self-borrow lifetime to *both* lifetimes
of the produced `Node`, artificially truncating its lifetime and
making this method ineffective in cases where the queried `Node`
was expected to match the lifetime of the originating `Document`’s
source data.