Use unconstrained lifetimes for the `attribute` functions (#17)
This commit adds unconstrained input lifetimes to the name parameters
passed in to the attribute lookup functions, so that the search text
does not need to be coupled to the lifetime of the XML parsed structures
being searched. Because the search text is only used for the duration of
the function call, and does not escape in the returned value, the search
term does not need to match the parsed lifetime.