Use Cow instead of Uri.
This allows us accessing Document/Node from multiple threads
which was not possible before due to Rc. And we cannot use Arc since
it will affect all the public API performance.

The downside is that when an XML has a lot of namespaces that should be normalized,
then it will be way slower and consume more memory.
But on an average XML it should be even faster.

Idealy, we should use some kind of sendable, read-only Rc.