Add benches for iter for some other libraries
Added descendant benches for xmltree and minidom as well as some
assertions to make sure the same things were being selected. Children
iteration benchmarks for left out for the other libraries because both
minidom and xmltree have Vec-backed child lists and we would only be
benchmarking how fast std Vec iterates, which is useless to us.