Remove first child from NodeData
Relying on the assumption that NodeId are assigned in document-order, we
can quickly find a Node's first child without storing it in NodeData. We
continue to store last child, since its relative placement to its parent
is variable. The last child field can continue to be used to determine
whether or not a Node has children.