Add `positions-extra-attr` feature for `Attribute` subranges.
To save memory this only uses two `u8`s per `Attribute`, but the tradeoff is that the qname and value ranges will be incorrect in a few extreme data cases, specifically if the attribute's qname has len > 255 or there are >254 spaces surrounding its equal sign.  Those cases seem unlikely enough in practice to justify not using more memory for this.