Drop tokenizer::Token intermediate data structure and XmlEvent trait
This simplifies the code base, especially the interaction between tokenizer
and parser. However, it also appears to imply that we have to drop the tokenizer
test which intimately rely on the intermediate token representation.

Further code clean-ups are possible, e.g. standardizing on methods on Context
instead of free functions, but this change aims for a minimal change while
we discuss this.