WIP: Add byte offset to `TextPos`.
I commented out the test requiring `size_of::<Error> <= 64` because the new data in `TextPos` increases it to 72.  That's obviously not the real solution, but I'm not sure what the feasible real solutions are.  Left a TODO.

I left a TODO about whether to include the byte offset in the `Display` impl for `TextPos`.

I added a new `TextPos::start` constructor for the not uncommon case where an error has no meaningful position and so just indicates the start of the data.